moved add sighting button to a lower level
