+++ src/components/ApplicationViews.js
@@ -9,161 +9,19 @@ import Welcome from './authentication/Welcome'
 // import SightingCard from './sighting/SightingCard'
 
 export default class ApplicationViews extends Component {
-  state = {
-    birds: [],
-    sightings: [],
-    users: [],
-    currentUserId:  +sessionStorage.getItem("userId") || +localStorage.getItem("userId"),
-    userName:"",
-    // date: "",
-    // location: "",
-    // birdId: "",
-    // summary: "",
-    editDate: "",
-    editLocation: "",
-    editBirdId: "",
-    editSummary: "",
-    editId: ""
-  };
+  isAuthenticated = () => (sessionStorage.getItem("userId") !== null || localStorage.getItem("userId") !== null)
 
-  isAuthenticated = () =>
-    sessionStorage.getItem("userId") !== null ||
-    localStorage.getItem("userId") !== null;
-
-  getAllUsers = () => APIManager.getAllEntries("users");
+  getAllUsers = () => APIManager.getAllEntries("users")
 
   getCurrentUser = () => {
-    const currentUser =
-      +sessionStorage.getItem("userId") || +localStorage.getItem("userId");
-    return currentUser;
-  };
-
-  componentDidMount() {
-    const newState = {};
-    APIManager.getAllEntries("birds").then(birds => {
-      this.setState({ birds: birds });
-    });
-
-    APIManager.getAllEntries("sightings", `?user_id=${this.state.currentUserId}&_sort=date&_order=asc&_expand=bird`)
-      .then(sightings => {
-        this.setState({ sightings: sightings });
-      })
-    
-    APIManager.getEntry("users", this.state.currentUserId)
-    .then((user) => {
-      this.setState({ userName: user.name })
-    })
-
-      .then(() => this.setState(newState));
+    const currentUser = +sessionStorage.getItem("userId") || +localStorage.getItem("userId")
+    return currentUser
   }
-
-  addSighting = sighting => {
-    return APIManager.addEntry("sightings", sighting)
-      .then(() =>
-        APIManager.getAllEntries(
-          "sightings",
-          `?user_id=${this.state.currentUserId}&_sort=date&_order=asc&_expand=bird`
-        )
-      )
-      .then(sightings =>
-        this.setState({
-          sightings: sightings
-        })
-      );
-  };
-
-  deleteSighting = id =>
-    APIManager.deleteEntry("sightings", id)
-      .then(() =>
-        APIManager.getAllEntries(
-          "sightings",
-          `?user_id=${this.state.currentUserId}&_sort=date&_order=asc&_expand=bird`
-        )
-      )
-      .then(sightings =>
-        this.setState({
-          sightings: sightings
-        })
-      );
-
-  editSighting = (editId, editSighting) =>
-    APIManager.editEntry("sightings", editId, editSighting)
-      .then(() =>
-        APIManager.getAllEntries(
-          "sightings",
-          `?user_id=${this.state.currentUserId}&_sort=date&_order=asc&_expand=bird`
-        )
-      )
-      .then(sightings =>
-        this.setState({
-          sightings: sightings
-        })
-      );
-
-  handleEditClick = (
-    editDate,
-    editLocation,
-    editBirdId,
-    editSummary,
-    editId
-  ) => {
-    this.setState({
-      editDate: editDate,
-      editLocation: editLocation,
-      editBirdId: editBirdId,
-      editSummary: editSummary,
-      editId: editId
-    });
-  };
-
-  handleFieldChange = evt => {
-    const stateToChange = {};
-    stateToChange[evt.target.id] = evt.target.value;
-    this.setState(stateToChange);
-  };
-
-  constructNewSighting = () => {
-    if (this.state.bird === "") {
-      window.alert("Please select a bird");
-    } else {
-      const sighting = {
-        date: this.state.date,
-        location: this.state.location,
-        birdId: this.state.birds.find(b => b.name === this.state.birdId).id,
-        summary: this.state.summary,
-        user_id: +this.state.currentUserId
-      };
-      console.log("sighting", sighting);
-      this.addSighting(sighting);
-    }
-  };
-
-  constructEditedSighting = () => {
-    const conditionBird = typeof this.state.editBirdId === "number";
-    const editSighting = {
-      date: this.state.editDate,
-      location: this.state.editLocation,
-      birdId: conditionBird
-        ? this.state.editBirdId
-        : this.state.birds.find(b => b.name === this.state.editBirdId).id,
-      summary: this.state.editSummary,
-      id: this.state.editId
-    };
-    console.log("edit sighting", editSighting);
-
-    this.editSighting(editSighting.id, editSighting);
-  };
 
   render() {
     return (
       <React.Fragment>
-        <NavBar
-          birds={this.state.birds}
-          sightings={this.state.sightings}
-          userName={this.state.userName}
-          handleFieldChange={this.handleFieldChange}
-          constructNewSighting={this.constructNewSighting}
-        />
+        <NavBar />
         <Route
           exact
           path="/"
@@ -172,13 +30,8 @@ export default class ApplicationViews extends Component {
               return (
                 <SightingList
                   {...props}
-                  birds={this.state.birds}
-                  sightings={this.state.sightings}
+                  getAllUsers={this.getAllUsers}
                   getCurrentUser={this.getCurrentUser}
-                  deleteSighting={this.deleteSighting}
-                  handleEditClick={this.handleEditClick}
-                  handleFieldChange={this.handleFieldChange}
-                  constructEditedSighting={this.constructEditedSighting}
                 />
               );
             } else {
